fixing makefiles etc

+++ cpp06/ex01/inc/Serializer.hpp
@@ -1,29 +1,23 @@
-#ifndef	SERIALIZER_HPP
+#ifndef SERIALIZER_HPP
 # define SERIALIZER_HPP
 # include <iostream>
-# include <cstdint>
-
 
 struct Data
 {
-	std::string	a;
-	std::string	b;
-	int	nbr;
-	int	nbr2;
+    std::string a;
+    std::string b;
+    int nbr;
+    int nbr2;
 };
-
 class Serializer
 {
 private:
-	Serializer();
-	Serializer(const Serializer& ref);
-	Serializer& operator=(const Serializer& ref);
-	~Serializer();
-	
+    Serializer();
+    Serializer(const Serializer& ref);
+    Serializer& operator=(const Serializer& ref);
+    ~Serializer();
 public:
-
-	static uintptr_t	serialize(Data* ptr);
-	static Data*		deserialize(uintptr_t raw);
+    static unsigned long    serialize(Data* ptr);
+    static Data*            deserialize(unsigned long raw);
 };
-
 #endif
